// src/generated/SideChain.ts
export interface SideChainDef {
    approveTrc20: (num: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'num' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    approveTrc721: (id: string, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'id' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    depositTrc: (functionSelector: string, num: number, fee: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'functionSelector' | 'num' | 'fee' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    depositTrc10: (tokenId: string, tokenValue: number, depositFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'tokenId' | 'tokenValue' | 'depositFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    depositTrc20: (num: number, depositFee: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'num' | 'depositFee' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    depositTrc721: (id: string, depositFee: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'id' | 'depositFee' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    depositTrx: (callValue: number, depositFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'callValue' | 'depositFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    injectFund: (num: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'num' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    mappingTrc: (trxHash: string, mappingFee: number, feeLimit: number, functionSelector: string, privateKey: string | null, callParams: CallParams<'trxHash' | 'mappingFee' | 'feeLimit' | 'functionSelector' | 'privateKey'>) => string | Promise<string>;
    mappingTrc20: (trxHash: string, mappingFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'trxHash' | 'mappingFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    mappingTrc721: (trxHash: string, mappingFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'trxHash' | 'mappingFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    retryDeposit: (nonce: number, retryDepositFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'nonce' | 'retryDepositFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    retryMapping: (nonce: number, retryMappingFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'nonce' | 'retryMappingFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    retryWithdraw: (nonce: number, retryWithdrawFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'nonce' | 'retryWithdrawFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    setChainId: (sideChainId: string, callParams: CallParams<'sideChainId'>) => void | Promise<void>;
    setMainGatewayAddress: (mainGatewayAddress: string, callParams: CallParams<'mainGatewayAddress'>) => void | Promise<void>;
    setSideGatewayAddress: (sideGatewayAddress: string, callParams: CallParams<'sideGatewayAddress'>) => void | Promise<void>;
    withdrawTrc: (functionSelector: string, num: number, withdrawFee: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'functionSelector' | 'num' | 'withdrawFee' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    withdrawTrc10: (tokenId: string, tokenValue: number, withdrawFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'tokenId' | 'tokenValue' | 'withdrawFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
    withdrawTrc20: (num: number, withdrawFee: number, feeLimit: number, contractAddress: string, privateKey: string | null, callParams: CallParams<'num' | 'withdrawFee' | 'feeLimit' | 'contractAddress' | 'privateKey'>) => string | Promise<string>;
    withdrawTrx: (callValue: number, withdrawFee: number, feeLimit: number, privateKey: string | null, callParams: CallParams<'callValue' | 'withdrawFee' | 'feeLimit' | 'privateKey'>) => string | Promise<string>;
}